:bug: fix: several bugs ole-web#2
modify some return value
change way of getting userinfo
change test of logout

+++ user/views.py
@@ -37,9 +37,9 @@ def login(req: HttpRequest):
             return request_success(return_data)
 
         body = json.loads(req.body.decode("utf-8"))
-        uname = require(body, "user_name", "string", err_msg="Missing or error type of name")
-        password = require(body, "password", err_msg="Missing or error type of password")
-        user: User = User.objects.filter(user_name=uname).first()
+        username = require(body, "user_name", "string", err_msg="invalid request", err_code=1005)
+        password = require(body, "password", "string", err_msg="invalid request", err_code=1005)
+        user: User = User.objects.filter(user_name=username).first()
         if not user:
             return request_failed(4, "wrong username or password", 400)
         else:
@@ -52,9 +52,10 @@ def login(req: HttpRequest):
                 response = request_success(return_data)
                 token = bcrypt.hashpw((str(user.user_id) + str(get_timestamp())).encode('utf-8'), bcrypt.gensalt())
                 while UserToken.objects.filter(token=token).exists():
-                    token = bcrypt.hashpw(((str(user.user_id) + str(get_timestamp())).encode('utf-8')), bcrypt.gensalt())
-                usertoken = UserToken(user=user, token=token)
-                usertoken.save()
+                    token = bcrypt.hashpw(((str(user.user_id) + str(get_timestamp())).encode('utf-8')),
+                                          bcrypt.gensalt())
+                user_token = UserToken(user=user, token=token)
+                user_token.save()
                 response.set_cookie("token", token)
                 return response
             else:
@@ -65,23 +66,27 @@ def login(req: HttpRequest):
 
 def logout(req: HttpRequest):
     if "token" in req.COOKIES:
-        response = request_success()
-        usertoken = UserToken.objects.get(token=req.COOKIES["token"])
-        usertoken.delete()
-        response.delete_cookie('token')
-        return response
-    else:
-        return request_success()
-
-
-def user_info(req: HttpRequest):
-    if req.method == "GET":
-        user = UserToken.objects.get(token=req.COOKIES["token"]).user
-        if not user:
-            return request_failed(1, "user not found")
+        if UserToken.objects.filter(token=req.COOKIES["token"]).exists():
+            response = request_success()
+            user_token = UserToken.objects.get(token=req.COOKIES["token"])
+            user_token.delete()
+            response.delete_cookie('token')
+            return response
         else:
-            return request_success(
-                {k: v for k, v in user.serialize().items() if k != "password"})
+            return request_failed(1001, "not_logged_in", 401)
+    else:
+        return request_failed(1001, "not_logged_in", 401)
+
+
+def user_info(req: HttpRequest, user_id: any):
+    user_id = require({"user_id": user_id}, "user_id", "int", err_msg="invalid request", err_code=1005)
+    if req.method == "GET":
+        if "token" not in req.COOKIES:
+            return request_failed(1001, "not_logged_in", 401)
+
+        user = User.objects.filter(user_id=user_id).first()
+        if not user:
+            return request_failed(8, "user does not exist", 404)
     else:
         return BAD_METHOD
